% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{prepare}
\alias{prepare}
\title{Dataset Preparation}
\usage{
prepare(auto = TRUE, ...)
}
\arguments{
\item{...}{Parameters to be passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything.
}
\description{
This function launches a Shiny app to help with the preparation
of a training dataset for YOLO.
}
\examples{
prepare()

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
