% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo.R
\name{remove_yolo}
\alias{remove_yolo}
\title{Remove YOLO}
\usage{
remove_yolo()
}
\value{
Nothing.
}
\description{
This function automates the removal of YOLO and all its Python
dependencies from your system.
}
\note{
The function will only remove the dedicated Python virtual environment from
your system. If Python was installed during the execution of
\code{\link[=install_yolo]{install_yolo()}}, it will not be removed.
}
\seealso{
\code{\link[=install_yolo]{install_yolo()}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
