% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{track}
\alias{track}
\title{Video Tracking}
\usage{
track(...)
}
\arguments{
\item{...}{Parameters to be passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything.
}
\description{
This function launches a Shiny app to help with the tracking of
objects in a video using a trained YOLO model.
}
\examples{
\dontrun{
track()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
