% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackRai-package.R
\docType{package}
\name{trackRai-package}
\alias{trackRai}
\alias{trackRai-package}
\title{trackRai: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/swarm-lab/trackRai}
  \item Report bugs at \url{https://github.com/swarm-lab/trackRai/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
