% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{train}
\alias{train}
\title{Model Training}
\usage{
train(...)
}
\arguments{
\item{...}{Parameters to be passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything.
}
\description{
This function launches a Shiny app to help with the training of
a YOLO model.
}
\examples{
\dontrun{
train()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
