% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{visualize}
\alias{visualize}
\title{Track Visualization}
\usage{
visualize(...)
}
\arguments{
\item{...}{Parameters to be passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything.
}
\description{
This function launches a Shiny app to help with the
visualization the tracking results obtained with \link{track}.
}
\examples{
\dontrun{
visualize()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
