% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo.R
\name{yolo_installed}
\alias{yolo_installed}
\title{Detect YOLO Installation}
\usage{
yolo_installed()
}
\value{
A logical indicating the presence or absence of YOLO.
}
\description{
This function detects whether YOLO is correctly installed on the
system.
}
\seealso{
\code{\link[=install_yolo]{install_yolo()}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
