% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo.R
\name{install_yolo}
\alias{install_yolo}
\title{Install and Update YOLO}
\usage{
install_yolo(python_version = "3.12.5")
}
\value{
If the installation/update completes successfully, a data frame
indicating the location of the YOLO installation and its version number.
}
\description{
This function automates the installation/updating of YOLO and
all its Python dependencies in a dedicated Python environment for use with
the \link{trackRai} apps.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
