% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackRai-package.R
\docType{package}
\name{trackRai-package}
\alias{trackRai-package}
\title{trackRai: Multi-Object Tracking With YOLO11}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

trackRai is a YOLO-based multi-object tracking software for R. It provides a series of Shiny apps for automating the preparation and training of YOLO11 models, for performing the tracking of multiple objects in a video, and for visualizing the results of the tracking process.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/swarm-lab/trackRai}
  \item \url{https://swarm-lab.github.io/trackRai/}
  \item Report bugs at \url{https://github.com/swarm-lab/trackRai/issues}
}

}
\author{
\strong{Maintainer}: Simon Garnier \email{garnier@njit.edu} (\href{https://orcid.org/0000-0002-3886-3974}{ORCID})

Other contributors:
\itemize{
  \item National Science Foundation, Award ID 2222418 [funder]
}

}
\keyword{internal}
